Add files via upload
The rent of a house depends on a lot of factors. With appropriate data and Machine Learning techniques, many real estate platforms find the housing options according to the customer’s budget. So, if you want to learn how to use Machine Learning to predict the rent of a house, this article is for you. In this article, I will take you through the task of House Rent Prediction with Machine Learning using Python.
The rent of a housing property depends on a lot of factors like:

number of bedrooms, hall, and kitchen
size of the property
the floor of the house
area type
area locality
City
furnishing status of the house
To build a house rent prediction system, we need data based on the factors affecting the rent of a housing property. I found a dataset from Kaggle which includes all the features we need.And posted the same in this repository